feat: att do JOptionpane
