+++ atividade_01/src/main/java/com/lista01/questao_04/Resposta_04.java
@@ -9,6 +9,7 @@ public class Resposta_04 {
 
 
         double primeiroNumero;
+
         do {
             System.out.println("Digite o primeiro numero (deve ser positivo e diferente de zero):");
             primeiroNumero = scanner.nextDouble();
