+++ atividade_01/src/main/java/com/lista01/questao_08/Resposta_08.java
@@ -14,28 +14,4 @@ public class Resposta_08 {
 
     }
 }
-class Pessoa{
-    private String nome;
-    private String sobreNome;
 
-    public Pessoa(String nome, String sobreNome){
-        this.nome = nome;
-        this.sobreNome = sobreNome;
-    }
-
-    public String getNome() {
-        return nome;
-    }
-
-    public void setNome(String nome) {
-        this.nome = nome;
-    }
-
-    public String getSobreNome() {
-        return sobreNome;
-    }
-
-    public void setSobreNome(String sobreNome) {
-        this.sobreNome = sobreNome;
-    }
-}
